requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
langchain-cohere
langchain
langchain-community
python-dotenv
jinja2==3.0.3
aiofiles
starlette>=0.14.0
python-multipart
torch